Added support for "chunk_size" parameter during queries.
This support has been added both to the config, and to the query opts.
The default value for chunk_size in config is nil, which causes no
functional change to how queries are executed. If chunk_size is set,
both the "chunked" and "chunk_size" parameters are set in the HTTP GET
and multi-line streaming JSON parsing occurs.

References:

- https://github.com/influxdata/influxdb/issues/6999
- https://github.com/influxdata/influxdb/issues/7000