Made explicit support for JSON streaming when chunking is enabled.
Now that https://github.com/influxdata/influxdb/issues/7000 has been
closed as the correct behaviour, and that multi-line JSON streaming is
the protocol when chunking is enabled, a specific parameter has been
plumbed through the query-to-http code path to enabled this feature
only when chunking is enabled.

This means the previous response handling is *unchanged*, removing the
risk of changing how it used to work and thus breaking anything. Missing
settings in the tests for the issue were also discovered and fixed.