Add support for configuring HTTP Proxy
The previous behavior was to adopt the global proxy setting through
the use of the Net::HTTP module.  This change introduces two new
configuration parameters that allow for manually specifying the HTTP
Proxy to use for the InfluxDB connection.

This allows for circumstances where it may be required to use a
specific proxy only for the InfluxDB connection.