Ensure workers can send data popped off the queue at shutdown
Previously, any data still in the queue would be handled by the main
thread. But, any data that a worker had already popped off the queue,
but not yet sent on the wire, would be lost.

This change pushes down the logic for stopping into the writer and
worker classes. The worker tells the threads to exit and waits for them
to do so.