Fix logging at level Logger::DEBUG
Previously setting the log level to Logger::DEBUG would have no effect
as the log level of the Logger object was not changed.

If you set:

    InfluxDB::Logging.log_level = Logger::DEBUG

InfluxDB::Logging::log? would return true, allowing the log statement to
proceed, but #log would not do anything because the Logger object was
still at its default from initialization, Logger::INFO.

By setting the log level directly on the Logger object and removing
::log? we allow the Logger object to determine if a log level needs to
be logged or not.

This allows debug-level log messages to be displayed.