Fix async persistent connections
With the addition of persistent connections we need to make a separate
InfluxDB::Client per worker as the persistent connection is attached to
the client.

Here this is done by copying the InfluxDB::Config object and creating a
new host queue (so threads don't have to communicate to share hosts) and
creating a new client without async enabled (so it will use HTTP or UDP
write methods).